refactor: modularize timeline component structure and enhance functionality
- Refactored the Timeline component to utilize custom hooks for managing timeline mode, navigation, media handling, and search functionality, improving code organization and reusability.
- Introduced a new TimelineView component to encapsulate rendering logic for different timeline modes, enhancing clarity and separation of concerns.
- Updated the useTimelineSearch hook to streamline search logic and improve focus management for search input.
- Enhanced the useTimelineNavigation hook to optimize item navigation and keyboard handling, improving user experience.
- Implemented useTimelineScroll and useTimelineMedia hooks for better scroll management and media visibility handling.
- Cleaned up unused code and improved type safety across components, ensuring better maintainability and performance.
